Make the engine build clean under -Wsign-compare (#1794)
Fourteen comparisons tested a signed value against a `size_t` or an
unsigned enum, so the signed side converted to unsigned. A guard
excludes a negative value at every one of them today, but the conversion
is what would hide the next one.

The object code is unchanged in `htscache`, `htsparse` and `punycode`,
and in every `htslib` function but `hts_log_vprint`. The other three
move a comparison from unsigned to signed or widen a loop index, so the
instructions change and the result does not. `hts_log_vprint` casts
`opt->debug` to int rather than retyping `level`, because an enum-typed
`level` reds `-Wswitch` on the switch below it.

The seven sites in `src/proxy/` follow separately.

Part of #1793

---------

+++ src/htsparse.c
@@ -1737,8 +1737,8 @@ int htsparse(htsmoduleStruct * str, htsmoduleStructExtended * stre) {
                                   // éventuellement remplire par des espaces pour avoir juste la taille
                                   while(strlen(tempo) < (size_t) (b - html))
                                     strcatbuff(tempo, " ");
-                                  // pas d'erreur?
-                                  if (strlen(tempo) == b - html) {
+                                  // no error?
+                                  if (strlen(tempo) == (size_t) (b - html)) {
                                     strncpy(modify, tempo, strlen(tempo)); // PAS d'octet nul à la fin!
                                     p = 0;      // DEVALIDER!!
                                     p_type = 0;
